🎉 Now I can request all my albums and parse it into a struct
This is the walking skeleton for the data retrieval. If _that_ works,
I can figure out the rest, like how to get specific images.

Of course now I need to clean up the mess a bit. Everything is
slapped into a giant main.

Also, the `Albums` struct right now suffers from "primitive obsession".
For lack of anything better, everything is a `String` when some
types really should be `Url`, etc. But we can get to that later.